requirements.txt
fastapi
uvicorn
requests
googlemaps
python-dotenv
google-genai
Pillow
langdetect
python-dotenv